Move PositionToCurrentController to JointLevelControllers
