+++ cmake/BipedalLocomotionFrameworkDependencies.cmake
@@ -293,6 +293,10 @@ framework_dependent_option(FRAMEWORK_COMPILE_ReducedModelControllers
   "Do you want to generate and compile the ReducedModelControllers?" ON
   "FRAMEWORK_USE_casadi;FRAMEWORK_COMPILE_System;FRAMEWORK_COMPILE_Contact;FRAMEWORK_COMPILE_Math;FRAMEWORK_COMPILE_CasadiConversions" OFF)
 
+framework_dependent_option(FRAMEWORK_COMPILE_JointLevelControllers
+  "Do you want to generate and compile the JointLevelControllers?" ON
+  "FRAMEWORK_COMPILE_System" OFF)  
+
 framework_dependent_option(FRAMEWORK_COMPILE_JointsGridPositionTrackingApplication
   "Compile joints-grid-position-tracking application?" ON
   "FRAMEWORK_COMPILE_YarpImplementation;FRAMEWORK_COMPILE_PYTHON_BINDINGS;FRAMEWORK_COMPILE_RobotInterface;FRAMEWORK_COMPILE_Math" OFF)
