+++ src/CMakeLists.txt
@@ -25,3 +25,4 @@ add_subdirectory(IK)
 add_subdirectory(SimplifiedModelControllers)
 add_subdirectory(ML)
 add_subdirectory(ReducedModelControllers)
+add_subdirectory(JointLevelControllers)
