+++ src/JointLevelControllers/tests/CMakeLists.txt
@@ -1,0 +1,5 @@
+# Copyright (C) 2025 Istituto Italiano di Tecnologia (IIT).
+
+if(FRAMEWORK_COMPILE_JointLevelControllers)
+
+endif()
